# backend/cmd/app/docs/swagger.yaml
basePath: /api
...
info:
  contact: {}
  description: Boiboi API Documentation.
  title: Boiboi Backend API
  version: "1.0"
paths:
...
  /admin/orders:
    get:
      consumes:
      - application/json
      description: Get a list of all orders with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                items:
                  additionalProperties: true
                  type: object
                type: array
              page:
                type: integer
              pageSize:
                type: integer
              totalCount:
                type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all orders
      tags:
      - Admin
  /admin/orders/{id}:
    get:
      consumes:
      - application/json
      description: Get a single order with full details by its ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an order by ID
      tags:
      - Admin
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"